requirements.txt
Flask==2.3.3
mysql-connector-python==8.0.34
